Covering more functionality with tests
