+++ tests/integration/TEC/Tickets/Commerce/Admin/Orders_PageTest.php
@@ -23,27 +23,6 @@ class Orders_PageTest extends \Codeception\TestCase\WPTestCase {
 	use With_Globals;
 
 	/**
-	 * Created orders.
-	 *
-	 * @var array
-	 */
-	protected $orders;
-
-	/**
-	 * Created tickets.
-	 *
-	 * @var array
-	 */
-	protected $tickets;
-
-	/**
-	 * Created event IDs.
-	 *
-	 * @var array
-	 */
-	protected $event_ids;
-
-	/**
 	 * @before
 	 */
 	public function set_up() {
@@ -103,6 +82,35 @@ class Orders_PageTest extends \Codeception\TestCase\WPTestCase {
 
 		$this->set_global_value( 'current_screen', WP_Screen::get( 'edit-' . Order::POSTTYPE ) );
 		$this->assertTrue( $orders_page->is_admin_orders_page() );
+	}
+
+	/**
+	 * @test
+	 */
+	public function it_should_locate_the_singular_order_page() {
+		$orders_page = new Orders_Page();
+
+		$this->set_global_value( 'current_screen', WP_Screen::get( 'edit' ) );
+		$this->assertFalse( $orders_page->is_admin_single_page() );
+
+		$this->set_global_value( 'current_screen', WP_Screen::get( Order::POSTTYPE ) );
+		$this->assertTrue( $orders_page->is_admin_single_page() );
+	}
+
+	/**
+	 * @test
+	 */
+	public function it_should_locate_the_orders_or_the_singular_order_page() {
+		$orders_page = new Orders_Page();
+
+		$this->set_global_value( 'current_screen', WP_Screen::get( 'edit' ) );
+		$this->assertFalse( $orders_page->is_admin_orders_page_or_admin_single_page() );
+
+		$this->set_global_value( 'current_screen', WP_Screen::get( Order::POSTTYPE ) );
+		$this->assertTrue( $orders_page->is_admin_orders_page_or_admin_single_page() );
+
+		$this->set_global_value( 'current_screen', WP_Screen::get( 'edit-' . Order::POSTTYPE ) );
+		$this->assertTrue( $orders_page->is_admin_orders_page_or_admin_single_page() );
 	}
 
 	/**
